refactor(sandbox-fc): split firecracker api client modules
