+++ crates/sandbox-fc/src/api/mod.rs
@@ -1,0 +1,10 @@
+mod client;
+mod http;
+
+#[cfg(test)]
+pub(crate) mod test_support;
+
+#[cfg(test)]
+mod tests;
+
+pub use client::{ApiClient, ApiError, BalloonStatistics};
